issue #6 (feat - api): dependency injection
