+++ src/services/impl/chapters_service.impl.ts
@@ -1,0 +1,36 @@
+import { IChapterDAO } from "../../dao/chapter_dao";
+import { IChapter } from "../../models/chapter";
+import { IChaptersService } from "../chapters_service";
+
+export class ChaptersService implements IChaptersService {
+     dao: IChapterDAO;
+
+     constructor(dao: IChapterDAO) {
+          this.dao = dao;
+     }
+
+     async getAll(): Promise<IChapter[]> {
+          return await this.dao.getAll();
+     }
+
+     async getById(id: number): Promise<IChapter | undefined> {
+          const chapter = await this.dao.getById(id);
+          return chapter;
+     }
+
+     async createChapter(chapter: IChapter): Promise<number | undefined> {
+          return await this.dao.create(chapter);
+     }
+
+     async deleteChapter(id: number) {
+          await this.dao.delete(id);
+     }
+
+     async updateChapter(
+          id: number,
+          newChapter: IChapter
+     ): Promise<IChapter | undefined> {
+          const chapter = await this.dao.update(id, newChapter);
+          return chapter;
+     }
+}
